renaming the product  - from rma10sk to RMA·Kalypso (the dot is written with Alt-pressed, Num-block 0-1-8-3, Alt-released) changes  - disable obvious code for 3D to make inoperative  - introduce 'implicit none' and 'only'-specifier for block usage in some subroutines  - start documenting the variables within the subroutines  - sorting variables definitions in logical blocks in order to be able to derive data types later on some bugfixes  - Moved reading of frequency for writing out iterations to the control file line C6. ('itefreq') (affects file.sub; input.sub)  - Deactivate option to define continuity lines in geometry *.2d-file; they can be only defined in the control file (affects rdkalyps.sub)  - derivative of velocity over time was not considered at the midside nodes in 1D polynomial approach (affects coef1dpoly.sub)  - replace imat(n)>0 durch imat(n)/=0 in getgeo.sub. Dry elements that are active for the calculation wouldn't be considered otherwise.  - small if-control change to increase robustness for writing out convergence feedback to the console in update.sub  - Increase the border to store relative changes for small models in update.sub  - Replacing common blocks IFU and swanf by module definitions.

+++ _import/trunk/fortran/RMA10S_TUHH/source/PolynomPositionSearch.F90
@@ -10,7 +10,7 @@ implicit none
 INTEGER :: FindPolynom
 !input
 !*****
-INTEGER, INTENT (IN) :: entries, node
+INTEGER (kind = 4), INTENT (IN) :: entries, node
 REAL (KIND = 8), INTENT (IN) :: rangearray(1:*)
 REAL (KIND = 8), INTENT (IN) :: depth
 real (kind = 8), intent (in) :: xcord, ycord
